+++ next.config.js
@@ -1,23 +1,11 @@
-const { createLoader } = require('simple-functional-loader');
-const rehypePrism = require('@mapbox/rehype-prism');
-const visit = require('unist-util-visit');
 const withBundleAnalyzer = require('@next/bundle-analyzer')({
   enabled: process.env.ANALYZE === 'true',
 });
 
-const tokenClassNames = {
-  tag: 'text-code-red',
-  'attr-name': 'text-code-yellow',
-  'attr-value': 'text-code-green',
-  deleted: 'text-code-red',
-  inserted: 'text-code-green',
-  punctuation: 'text-code-white',
-  keyword: 'text-code-purple',
-  string: 'text-code-green',
-  function: 'text-code-blue',
-  boolean: 'text-code-red',
-  comment: 'text-gray-400 italic',
-};
+const createMDXLoader = require('./lib/mdx-loader');
+const createMDXPreviewLoader = require('./lib/mdx-preview-loader');
+const createPostLoader = require('./lib/post-loader');
+const createRSSLoader = require('./lib/rss-loader');
 
 module.exports = withBundleAnalyzer({
   pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'mdx'],
@@ -38,73 +26,21 @@ module.exports = withBundleAnalyzer({
       ],
     });
 
-    const mdx = [
-      options.defaultLoaders.babel,
-      {
-        loader: '@mdx-js/loader',
-        options: {
-          rehypePlugins: [
-            rehypePrism,
-            () => {
-              return (tree) => {
-                visit(tree, 'element', (node, index, parent) => {
-                  let [token, type] = node.properties.className || [];
-                  if (token === 'token') {
-                    node.properties.className = [tokenClassNames[type]];
-                  }
-                });
-              };
-            },
-          ],
-        },
-      },
-    ];
+    const mdx = createMDXLoader(options);
 
     config.module.rules.push({
       test: /\.mdx$/,
       oneOf: [
         {
           resourceQuery: /preview/,
-          use: [
-            ...mdx,
-            createLoader(function (src) {
-              if (src.includes('<!--more-->')) {
-                const [preview] = src.split('<!--more-->');
-                return this.callback(null, preview);
-              }
-
-              const [preview] = src.split('<!--/excerpt-->');
-              return this.callback(null, preview.replace('<!--excerpt-->', ''));
-            }),
-          ],
+          use: [...mdx, createMDXPreviewLoader(options)],
         },
         {
           resourceQuery: /rss/,
-          use: [
-            ...mdx,
-            createLoader(function (src) {
-              return this.callback(null, src);
-            }),
-          ],
+          use: [...mdx, createRSSLoader()],
         },
         {
-          use: [
-            ...mdx,
-            createLoader(function (src) {
-              const content = [
-                'import Post from "~/components/Post"',
-                'export { getStaticProps } from "~/getStaticProps"',
-                src,
-                'export default Post',
-              ].join('\n');
-
-              if (content.includes('<!--more-->')) {
-                return this.callback(null, content.split('<!--more-->').join('\n'));
-              }
-
-              return this.callback(null, content.replace(/<!--excerpt-->.*<!--\/excerpt-->/s, ''));
-            }),
-          ],
+          use: [...mdx, createPostLoader()],
         },
       ],
     });
