+++ package.json
@@ -32,15 +32,16 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "styled-jsx-plugin-postcss": "^4.0.1",
+    "simple-functional-loader": "^1.2.1",
     "tailwindcss": "^1.9.6",
-    "tinytime": "^0.2.6"
+    "tinytime": "^0.2.6",
+    "unist-util-visit": "^2.0.3"
   },
   "license": "ISC",
   "devDependencies": {
     "@next/bundle-analyzer": "^10.0.6",
     "@testing-library/dom": "^7.29.4",
     "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.2",
     "@testing-library/react": "^11.2.5",
     "@types/jest": "^26.0.20",
     "@types/node": "^14.14.25",
