+++ src/components/MDX/index.tsx
@@ -1,0 +1,10 @@
+import Image from 'next/image';
+
+import Code from './Code';
+import Pre from './Pre';
+
+export default {
+  pre: Pre,
+  'pre.code': Code,
+  Image,
+};
